Change name of person class to player
